- changed datatype init

+++ utils.py
@@ -304,15 +304,15 @@ def get_mask(mask_name, dilate=0, dtype=np.uint8):
 
 def get_normalised_image(image_name, mask, dtype=np.float32, masked=False):
     mask_bin = mask.astype(np.bool)
-    image = load_nii(image_name).get_data()
+    image = load_nii(image_name).get_data().astype(dtype)
     image_mu = np.mean(image[mask_bin])
     image_sigma = np.std(image[mask_bin])
     norm_image = (image - image_mu) / image_sigma
 
     if masked:
-        output = norm_image.astype(dtype) * mask_bin.astype(dtype)
+        output = norm_image * mask_bin.astype(dtype)
     else:
-        output = norm_image.astype(dtype)
+        output = norm_image
 
     return output
